Move assettasks to pkg/assets
