+++ pkg/assets/BUILD.bazel
@@ -2,10 +2,18 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["builder.go"],
+    srcs = [
+        "builder.go",
+        "copy.go",
+        "copyfile.go",
+        "copyimage.go",
+        "docker_api.go",
+        "docker_cli.go",
+    ],
     importpath = "k8s.io/kops/pkg/assets",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/acls:go_default_library",
         "//pkg/apis/kops:go_default_library",
         "//pkg/apis/kops/util:go_default_library",
         "//pkg/kubemanifest:go_default_library",
@@ -14,6 +22,9 @@ go_library(
         "//util/pkg/mirrors:go_default_library",
         "//util/pkg/vfs:go_default_library",
         "//vendor/github.com/blang/semver/v4:go_default_library",
+        "//vendor/github.com/docker/docker/api/types:go_default_library",
+        "//vendor/github.com/docker/docker/api/types/filters:go_default_library",
+        "//vendor/github.com/docker/docker/client:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/wait:go_default_library",
         "//vendor/k8s.io/klog/v2:go_default_library",
     ],
@@ -21,7 +32,10 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["builder_test.go"],
+    srcs = [
+        "builder_test.go",
+        "copyfile_test.go",
+    ],
     data = glob(["testdata/**"]),
     embed = [":go_default_library"],
     deps = [
