+++ pkg/assets/docker_cli.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package assettasks
+package assets
 
 import (
 	"fmt"
