--- upup/pkg/fi/assettasks/BUILD.bazel
@@ -1,31 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "copy.go",
-        "copyfile.go",
-        "copyimage.go",
-        "docker_api.go",
-        "docker_cli.go",
-    ],
-    importpath = "k8s.io/kops/upup/pkg/fi/assettasks",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/acls:go_default_library",
-        "//pkg/apis/kops:go_default_library",
-        "//pkg/assets:go_default_library",
-        "//util/pkg/hashing:go_default_library",
-        "//util/pkg/vfs:go_default_library",
-        "//vendor/github.com/docker/docker/api/types:go_default_library",
-        "//vendor/github.com/docker/docker/api/types/filters:go_default_library",
-        "//vendor/github.com/docker/docker/client:go_default_library",
-        "//vendor/k8s.io/klog/v2:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["copyfile_test.go"],
-    embed = [":go_default_library"],
-)
